Implement unit test platform, add some tests
