+++ package.json
@@ -1,0 +1,34 @@
+{
+  "name": "witchcraft",
+  "version": "2.2.0",
+  "description": "![Witchcraft](docs/title.png)",
+  "main": "index.js",
+  "directories": {
+    "doc": "docs"
+  },
+  "scripts": {
+    "test": "nyc mocha"
+  },
+  "nyc": {
+    "all": true,
+    "include": ["chrome-extension/*.js"]
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/luciopaiva/witchcraft.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/luciopaiva/witchcraft/issues"
+  },
+  "homepage": "https://github.com/luciopaiva/witchcraft#readme",
+  "dependencies": {},
+  "devDependencies": {
+    "mocha": "^5.2.0",
+    "nyc": "^13.0.1",
+    "sinon": "^6.1.5",
+    "sinon-chrome": "^2.3.2"
+  }
+}
